+++ resources/models/reservation.py
@@ -37,9 +37,6 @@ RESERVATION_EXTRA_FIELDS = ('reserver_name', 'reserver_phone_number', 'reserver_
                             'number_of_participants', 'participants', 'reserver_email_address', 'host_name'
                             )
 
-# after this time reservations in state "waiting_for_payment" without an order will be cancelled
-ORDER_WAITING_TIME = datetime.timedelta(seconds=30)
-
 
 class ReservationQuerySet(models.QuerySet):
     def current(self):
@@ -515,18 +512,6 @@ class Reservation(ModifiableModel):
                 self.access_code = generate_access_code(access_code_type)
 
         return super().save(*args, **kwargs)
-
-    @classmethod
-    def cancel_too_old_reservations_without_orders(cls):
-        earliest_allowed = now() - ORDER_WAITING_TIME
-        reservations_to_cancel = cls.objects.filter(
-            state=cls.WAITING_FOR_PAYMENT,
-            order=None,
-            created_at__lt=earliest_allowed,
-        )
-        for reservation in reservations_to_cancel:
-            reservation.set_state(cls.CANCELLED, None)
-        return reservations_to_cancel.count()
 
 
 class ReservationMetadataField(models.Model):
